dev-java/oracle-jdk-bin: Fix -alsa for 11, make -javafx wildcard wilder
Java 11 has no libjsoundalsa.*, check for file before removal with +alsa
Make the wildcard for removal of lib/* files for -javafx wilder to match
more files for removal. Removes files that were omitted from removal.

Package-Manager: Portage-2.3.24, Repoman-2.3.6

+++ dev-java/oracle-jdk-bin/oracle-jdk-bin-9999.ebuild
@@ -136,7 +136,7 @@ src_prepare() {
 	rm lib/libavplugin* || die "Failed to remove libavplugin*.so"
 	rm lib/fontconfig.* || die "Failed to remove fontconfig.*"
 
-	if ! use alsa ; then
+	if ! use alsa && [[ -f lib/libjsoundalsa.* ]]; then
 		rm lib/libjsoundalsa.* \
 			|| die "Failed to remove unwanted alsa support"
 	fi
@@ -152,7 +152,7 @@ src_prepare() {
 
 	if ! use javafx ; then
 		rm jmods/javafx*  \
-			lib/lib*{decora,fx,glass,gstreamer,prism}* \
+			lib/*{decora,fx,glass,gstreamer,prism}* \
 			|| die "Failed to remove unwanted JavaFX support"
 	else
 		if ! use gtk2 ; then
